Extract common DiffCallback code to be shared between Section and GroupAdapter - Also create GroupUtils to put other shared functionality there
